pyproject.toml
[project]
name = "color"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
